# packages/backend/public/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"IKeyPayload": {
				"properties": {
					"keyId": {
						"type": "string"
					},
					"userId": {
						"type": "string"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"keyId",
					"userId",
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"INewKey": {
				"properties": {
					"payload": {
						"$ref": "#/components/schemas/IKeyPayload"
					},
					"key": {
						"type": "string"
					}
				},
				"required": [
					"payload",
					"key"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"key": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header"
			},
			"token": {
				"type": "http",
				"scheme": "bearer"
			}
		}
	},
	"info": {
		"title": "backend",
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/v1/auth/key": {
			"get": {
				"operationId": "GetKeys",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IKeyPayload"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [
					{
						"token": []
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "CreateKey",
				"responses": {
					"201": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/INewKey"
								}
							}
						}
					}
				},
				"security": [
					{
						"token": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "string"
							}
						}
					}
				}
			}
		},
		"/v1/auth/key/{id}": {
			"get": {
				"operationId": "GetKey",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IKeyPayload"
								}
							}
						}
					},
					"404": {
						"description": ""
					}
				},
				"security": [
					{
						"token": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteKey",
				"responses": {
					"204": {
						"description": ""
					},
					"404": {
						"description": ""
					}
				},
				"security": [
					{
						"token": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/v1/ping": {
			"get": {
				"operationId": "GetMessage",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}